# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyCore API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes
      parameters:
        - name: subjectId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    put:
      operationId: updateNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Flashcard Decks ────────────────────────────────────────────────────────
  /flashcard-decks:
    get:
      operationId: listFlashcardDecks
      tags: [flashcards]
      summary: List all flashcard decks
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardDeck"
    post:
      operationId: createFlashcardDeck
      tags: [flashcards]
      summary: Create a flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlashcardDeckBody"
      responses:
        "201":
          description: Deck created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
  /flashcard-decks/{id}:
    get:
      operationId: getFlashcardDeck
      tags: [flashcards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deck with cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeckDetail"
    delete:
      operationId: deleteFlashcardDeck
      tags: [flashcards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /flashcard-decks/{id}/cards:
    get:
      operationId: listFlashcards
      tags: [flashcards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cards in deck
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
    post:
      operationId: addFlashcard
      tags: [flashcards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlashcardBody"
      responses:
        "201":
          description: Card added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"

  # ── Glossary ───────────────────────────────────────────────────────────────
...
  /sets:
    get:
      operationId: listSets
      tags: [sets]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of study sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudySet"
    post:
      operationId: createSet
      tags: [sets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySetBody"
      responses:
        "201":
          description: Set created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
  /sets/{id}:
    get:
      operationId: getSet
      tags: [sets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Study set detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
    delete:
      operationId: deleteSet
      tags: [sets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Global Search ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Note:
      type: object
      required: [id, title, content, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        diagramUrl:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateNoteBody:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        diagramUrl:
          type: string
    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        diagramUrl:
          type: string
        bulletPoints:
          type: array
          items:
            type: string

    # ── Flashcards ────────────────────────────────────────────────────────────
    Flashcard:
      type: object
      required: [id, deckId, front, back]
      properties:
        id:
          type: string
        deckId:
          type: string
        front:
          type: string
        back:
          type: string
        hint:
          type: string
    FlashcardDeck:
      type: object
      required: [id, title, cardCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        cardCount:
          type: integer
        subject:
          type: string
        createdAt:
          type: string
    FlashcardDeckDetail:
      type: object
      required: [id, title, cards, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        createdAt:
          type: string
    CreateFlashcardDeckBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
    CreateFlashcardBody:
      type: object
      required: [front, back]
      properties:
        front:
          type: string
        back:
          type: string
        hint:
          type: string

    # ── Glossary ──────────────────────────────────────────────────────────────
...
    StudySet:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        termCount:
          type: integer
        createdAt:
          type: string
    CreateStudySetBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string

    # ── Search ─────────────────────────────────────────────────────────────────
...